requirements.txt
beautifulsoup4==4.12.2
pandas==2.0.3
requests==2.31.0
readability-lxml==0.8.1
nltk==3.8.1
textstat==0.7.3
textblob==0.17.1
tqdm==4.66.1
openpyxl==3.1.2